[Core] Move memory monitors into common/monitor (#64491)
## Description
This PR moves all monitors and their related tests into
`src/ray/common/monitors` for organizational purposes. Previously, we
only had a single type of monitor. Now we have helpers dedicated for
monitoring both cpu and memory. And there are multiple types of memory
monitors. Thus, we move them for clarity.

## Related issues

## Additional information
