+++ src/ray/common/BUILD.bazel
@@ -90,177 +90,6 @@ ray_cc_library(
 )
 
 ray_cc_library(
-    name = "memory_monitor_test_fixture",
-    srcs = ["memory_monitor_test_fixture.cc"],
-    hdrs = ["memory_monitor_test_fixture.h"],
-    target_compatible_with = [
-        "@platforms//os:linux",
-    ],
-    visibility = [
-        "//src/ray/common/tests:__pkg__",
-    ],
-    deps = [
-        ":memory_monitor_utils",
-        "//src/ray/common/cgroup2:cgroup_test_utils",
-        "//src/ray/util:logging",
-        "@com_google_googletest//:gtest",
-    ],
-)
-
-ray_cc_library(
-    name = "memory_monitor_interface",
-    hdrs = [
-        "memory_monitor_interface.h",
-    ],
-    visibility = [
-        ":__subpackages__",
-        "//src/ray/raylet:__subpackages__",
-    ],
-    deps = [
-        "//src/ray/util:compat",
-        "@com_google_absl//absl/container:flat_hash_map",
-    ],
-)
-
-ray_cc_library(
-    name = "memory_monitor_utils",
-    srcs = [
-        "memory_monitor_utils.cc",
-    ],
-    hdrs = [
-        "memory_monitor_utils.h",
-    ],
-    visibility = [
-        ":__subpackages__",
-        "//src/ray/raylet:__subpackages__",
-    ],
-    deps = [
-        ":memory_monitor_interface",
-        ":ray_config",
-        "//src/ray/common/cgroup2:cgroup_manager_interface",
-        "//src/ray/util:compat",
-        "//src/ray/util:logging",
-        "@boost//:algorithm",
-        "@com_google_absl//absl/strings",
-        "@com_google_absl//absl/strings:str_format",
-        "@com_google_googletest//:gtest_prod",
-    ],
-)
-
-ray_cc_library(
-    name = "threshold_memory_monitor",
-    srcs = [
-        "threshold_memory_monitor.cc",
-    ],
-    hdrs = [
-        "threshold_memory_monitor.h",
-    ],
-    target_compatible_with = [
-        "@platforms//os:linux",
-    ],
-    visibility = [":__subpackages__"],
-    deps = [
-        ":memory_monitor_interface",
-        ":memory_monitor_utils",
-        "//src/ray/asio:periodical_runner",
-        "//src/ray/util:logging",
-        "//src/ray/util:thread_utils",
-        "@com_google_absl//absl/strings:str_format",
-    ],
-)
-
-ray_cc_library(
-    name = "noop_memory_monitor",
-    hdrs = [
-        "noop_memory_monitor.h",
-    ],
-    visibility = [":__subpackages__"],
-    deps = [
-        ":memory_monitor_interface",
-    ],
-)
-
-ray_cc_library(
-    name = "event_memory_monitor",
-    srcs = [
-        "event_memory_monitor.cc",
-    ],
-    hdrs = [
-        "event_memory_monitor.h",
-    ],
-    target_compatible_with = [
-        "@platforms//os:linux",
-    ],
-    visibility = [":__subpackages__"],
-    deps = [
-        ":memory_monitor_interface",
-        ":memory_monitor_utils",
-        ":status",
-        "//src/ray/util:logging",
-        "//src/ray/util:thread_utils",
-        "@boost//:algorithm",
-        "@com_google_absl//absl/strings:str_format",
-    ],
-)
-
-ray_cc_library(
-    name = "pressure_memory_monitor",
-    srcs = [
-        "pressure_memory_monitor.cc",
-    ],
-    hdrs = [
-        "pressure_memory_monitor.h",
-    ],
-    target_compatible_with = [
-        "@platforms//os:linux",
-    ],
-    visibility = [":__subpackages__"],
-    deps = [
-        ":memory_monitor_interface",
-        ":memory_monitor_utils",
-        ":status_or",
-        "//src/ray/util:logging",
-        "//src/ray/util:thread_utils",
-        "@com_google_absl//absl/strings:str_format",
-    ],
-)
-
-ray_cc_library(
-    name = "memory_monitor_factory",
-    srcs = select({
-        "//bazel:is_linux": [
-            "multi_memory_monitor_factory.cc",
-        ],
-        "//conditions:default": [
-            "noop_memory_monitor_factory.cc",
-        ],
-    }),
-    hdrs = [
-        "memory_monitor_factory.h",
-    ],
-    visibility = [
-        ":__subpackages__",
-        "//src/ray/raylet:__subpackages__",
-    ],
-    deps = [
-        ":memory_monitor_interface",
-        "//src/ray/common/cgroup2:cgroup_manager_interface",
-    ] + select({
-        "//bazel:is_linux": [
-            ":event_memory_monitor",
-            ":memory_monitor_utils",
-            ":noop_memory_monitor",
-            ":ray_config",
-            ":threshold_memory_monitor",
-            "//src/ray/util:logging",
-        ],
-        "//conditions:default": [
-            ":noop_memory_monitor",
-        ],
-    }),
-)
-
-ray_cc_library(
     name = "file_system_monitor",
     srcs = ["file_system_monitor.cc"],
     hdrs = ["file_system_monitor.h"],
