+++ src/ray/common/cgroup2/BUILD.bazel
@@ -26,9 +26,9 @@ ray_cc_library(
             ":cgroup_driver_interface",
             ":cgroup_manager",
             ":sysfs_cgroup_driver",
-            "//src/ray/common:memory_monitor_utils",
             "//src/ray/common:status",
             "//src/ray/common:status_or",
+            "//src/ray/common/monitors:memory_monitor_utils",
             "@com_google_absl//absl/strings",
         ],
         "//conditions:default": [],
@@ -60,7 +60,7 @@ ray_cc_library(
     ],
     visibility = [
         ":__subpackages__",
-        "//src/ray/common/tests:__pkg__",
+        "//src/ray/common/monitors/tests:__pkg__",
     ],
     deps = [
         ":cgroup_driver_interface",
@@ -79,7 +79,7 @@ ray_cc_library(
     ],
     visibility = [
         ":__subpackages__",
-        "//src/ray/common/tests:__pkg__",
+        "//src/ray/common/monitors/tests:__pkg__",
         "//src/ray/raylet/tests:__pkg__",
     ],
     deps = [
@@ -126,7 +126,7 @@ ray_cc_library(
     ],
     visibility = [
         ":__subpackages__",
-        "//src/ray/common/tests:__pkg__",
+        "//src/ray/common/monitors/tests:__pkg__",
     ],
     deps = [
         ":cgroup_driver_interface",
@@ -144,8 +144,8 @@ ray_cc_library(
     visibility = [
         ":__subpackages__",
         "//src/ray/common:__pkg__",
+        "//src/ray/common/monitors:__pkg__",
         "//src/ray/common/monitors/tests:__pkg__",
-        "//src/ray/common/tests:__pkg__",
     ],
     deps = [
         "//src/ray/common:id",
@@ -163,7 +163,7 @@ ray_cc_library(
     ],
     visibility = [
         ":__subpackages__",
-        "//src/ray/common/tests:__pkg__",
+        "//src/ray/common/monitors/tests:__pkg__",
     ],
     deps = [
         ":cgroup_manager_interface",
