+++ src/ray/common/cgroup2/linux_cgroup_manager_factory.cc
@@ -28,7 +28,7 @@
 #include "ray/common/cgroup2/cgroup_manager_interface.h"
 #include "ray/common/cgroup2/noop_cgroup_manager.h"
 #include "ray/common/cgroup2/sysfs_cgroup_driver.h"
-#include "ray/common/memory_monitor_utils.h"
+#include "ray/common/monitors/memory_monitor_utils.h"
 #include "ray/common/ray_config.h"
 
 namespace ray {
