+++ src/ray/common/monitors/memory_monitor_utils.h
@@ -24,7 +24,7 @@
 #include <vector>
 
 #include "ray/common/cgroup2/cgroup_manager_interface.h"
-#include "ray/common/memory_monitor_interface.h"
+#include "ray/common/monitors/memory_monitor_interface.h"
 #include "ray/common/status_or.h"
 #include "ray/util/compat.h"
