+++ src/ray/common/monitors/multi_memory_monitor_factory.cc
@@ -15,13 +15,13 @@
 #include <memory>
 #include <vector>
 
-#include "ray/common/event_memory_monitor.h"
-#include "ray/common/memory_monitor_factory.h"
-#include "ray/common/memory_monitor_interface.h"
-#include "ray/common/memory_monitor_utils.h"
-#include "ray/common/noop_memory_monitor.h"
+#include "ray/common/monitors/event_memory_monitor.h"
+#include "ray/common/monitors/memory_monitor_factory.h"
+#include "ray/common/monitors/memory_monitor_interface.h"
+#include "ray/common/monitors/memory_monitor_utils.h"
+#include "ray/common/monitors/noop_memory_monitor.h"
+#include "ray/common/monitors/threshold_memory_monitor.h"
 #include "ray/common/ray_config.h"
-#include "ray/common/threshold_memory_monitor.h"
 #include "ray/util/logging.h"
 
 namespace ray {
