+++ src/ray/common/monitors/pressure_memory_monitor.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "ray/common/pressure_memory_monitor.h"
+#include "ray/common/monitors/pressure_memory_monitor.h"
 
 #include <fcntl.h>
 #include <sys/eventfd.h>
@@ -22,7 +22,7 @@
 #include <filesystem>
 #include <string>
 
-#include "ray/common/memory_monitor_utils.h"
+#include "ray/common/monitors/memory_monitor_utils.h"
 #include "ray/util/logging.h"
 #include "ray/util/thread_utils.h"
