+++ src/ray/common/monitors/threshold_memory_monitor.cc
@@ -12,10 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "ray/common/threshold_memory_monitor.h"
+#include "ray/common/monitors/threshold_memory_monitor.h"
 
 #include "absl/strings/str_format.h"
-#include "ray/common/memory_monitor_utils.h"
+#include "ray/common/monitors/memory_monitor_utils.h"
 #include "ray/util/logging.h"
 #include "ray/util/thread_utils.h"
