+++ src/ray/common/tests/BUILD.bazel
@@ -86,121 +86,6 @@ ray_cc_test(
 )
 
 ray_cc_test(
-    name = "memory_monitor_utils_test",
-    size = "small",
-    srcs = [
-        "memory_monitor_utils_test.cc",
-    ],
-    tags = [
-        "no_windows",
-        "team:core",
-    ],
-    target_compatible_with = [
-        "@platforms//os:linux",
-    ],
-    deps = [
-        "//src/ray/common:id",
-        "//src/ray/common:memory_monitor_test_fixture",
-        "//src/ray/common:memory_monitor_utils",
-        "//src/ray/common:ray_config",
-        "//src/ray/common/cgroup2:cgroup_manager",
-        "//src/ray/common/cgroup2:cgroup_test_utils",
-        "//src/ray/common/cgroup2:fake_cgroup_driver",
-        "//src/ray/common/cgroup2:noop_cgroup_manager",
-        "//src/ray/util:process",
-        "@boost//:filesystem",
-        "@com_google_absl//absl/container:flat_hash_map",
-    ],
-)
-
-ray_cc_test(
-    name = "threshold_memory_monitor_test",
-    size = "small",
-    srcs = [
-        "threshold_memory_monitor_test.cc",
-    ],
-    tags = [
-        "no_windows",
-        "team:core",
-    ],
-    target_compatible_with = [
-        "@platforms//os:linux",
-    ],
-    deps = [
-        "//src/ray/common:memory_monitor_interface",
-        "//src/ray/common:memory_monitor_test_fixture",
-        "//src/ray/common:memory_monitor_utils",
-        "//src/ray/common:threshold_memory_monitor",
-        "//src/ray/common/cgroup2:noop_cgroup_manager",
-        "@boost//:thread",
-    ],
-)
-
-ray_cc_test(
-    name = "event_memory_monitor_test",
-    size = "small",
-    srcs = [
-        "event_memory_monitor_test.cc",
-    ],
-    tags = [
-        "no_windows",
-        "team:core",
-    ],
-    target_compatible_with = [
-        "@platforms//os:linux",
-    ],
-    deps = [
-        "//src/ray/common:event_memory_monitor",
-        "//src/ray/common/cgroup2:cgroup_test_utils",
-        "//src/ray/util:logging",
-        "@boost//:thread",
-    ],
-)
-
-ray_cc_test(
-    name = "pressure_memory_monitor_test",
-    size = "small",
-    srcs = [
-        "pressure_memory_monitor_test.cc",
-    ],
-    tags = [
-        "no_windows",
-        "team:core",
-    ],
-    target_compatible_with = [
-        "@platforms//os:linux",
-    ],
-    deps = [
-        "//src/ray/common:pressure_memory_monitor",
-        "//src/ray/common/cgroup2:cgroup_test_utils",
-        "//src/ray/util:logging",
-        "@boost//:thread",
-    ],
-)
-
-ray_cc_test(
-    name = "memory_monitor_factory_test",
-    size = "small",
-    srcs = [
-        "memory_monitor_factory_test.cc",
-    ],
-    tags = [
-        "no_windows",
-        "team:core",
-    ],
-    target_compatible_with = [
-        "@platforms//os:linux",
-    ],
-    deps = [
-        "//src/ray/common:event_memory_monitor",
-        "//src/ray/common:memory_monitor_factory",
-        "//src/ray/common:memory_monitor_interface",
-        "//src/ray/common:threshold_memory_monitor",
-        "//src/ray/common/cgroup2:fake_cgroup_manager",
-    ],
-)
-
-ray_cc_test(
     name = "grpc_util_test",
     size = "small",
     srcs = [
