+++ src/ray/raylet/worker_killing_policy_by_time.cc
@@ -25,7 +25,7 @@
 #include "absl/strings/str_join.h"
 #include "absl/time/time.h"
 #include "ray/common/lease/lease.h"
-#include "ray/common/memory_monitor_utils.h"
+#include "ray/common/monitors/memory_monitor_utils.h"
 #include "ray/util/compat.h"
 
 namespace ray {
