+++ src/ray/raylet/worker_killing_policy_factory.cc
@@ -17,7 +17,7 @@
 #include <algorithm>
 #include <memory>
 
-#include "ray/common/memory_monitor_utils.h"
+#include "ray/common/monitors/memory_monitor_utils.h"
 #include "ray/common/ray_config.h"
 #include "ray/raylet/worker_killing_policy_by_time.h"
 #include "ray/raylet/worker_killing_policy_group_by_owner.h"
