+++ src/ray/raylet/worker_killing_policy_interface.h
@@ -18,7 +18,7 @@
 #include <utility>
 #include <vector>
 
-#include "ray/common/memory_monitor_interface.h"
+#include "ray/common/monitors/memory_monitor_interface.h"
 #include "ray/raylet/worker_interface.h"
 
 namespace ray {
